pyproject.toml
[project]
name = "credly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.41.3",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
]
